pyproject.toml
[tool.poetry]
name = "warehouse-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "warehouse_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-rest-framework = "^0.1.0"
pillow = "^10.0.0"
django-import-export = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
